로그아웃시 브라우저 쿠키삭제 완료

+++ JoinUs/member_app/views.py
@@ -3,23 +3,34 @@ from .models import User
 from django.http import JsonResponse, HttpResponseRedirect
 from django.urls import reverse
 from django.apps import apps  # 다른 어플 모델 참조
-# Create your views here.
+# 장고 로그아웃 관리
+from django.contrib import auth
+
+# 회원 가입 페이지로 이동
 
 
-def signin(request):
-    return render(request, 'member_app/signin.html')
+def signup(request):
+    if request.session.get('user'):
+        return HttpResponseRedirect(reverse('index'))
+    else:
+        return render(request, 'member_app/signup.html')
 
 
 # 회원 가입
 def register(request):
-    user_email = request.POST['user_email']
-    user_nickname = request.POST['user_nickname']
-    user_pw = request.POST['user_pw']
-    new_user = User(user_email=user_email,
-                    user_nickname=user_nickname, user_pw=user_pw)
-    new_user.save()
-    # 다른 view 메서드 바로 참조 가능(url아님)
-    return HttpResponseRedirect(reverse("signupPage"))
+
+    if request.method == 'POST':
+        user_email = request.POST['user_email']
+        user_nickname = request.POST['user_nickname']
+        user_pw = request.POST['user_pw']
+        new_user = User(user_email=user_email,
+                        user_nickname=user_nickname, user_pw=user_pw)
+        new_user.save()
+        # 다른 view 메서드 바로 참조 가능(url아님)
+        return HttpResponseRedirect(reverse("signinPage"))
+    elif request.method == 'GET':
+        res_data = {'error': '잘못된 경로입니다'}
+        render(request, 'member_app/signup.html', res_data)
 
 
 # 이메일 중복 체크
@@ -50,10 +61,11 @@ def nicknameCheck(request):
     return JsonResponse(result)
 
 
-# 회원 페이지에서 닉네임 변경
+# 회원 페이지에서 닉네임 변경(Ajax)
 def updateNickname(request):
     user_id = request.session.get('user_id')
     same = User.objects.get(u_id=user_id)
+    # 기존에 등록된 유저 닉네임과 같지 않으면
     if request.GET['user_nickname'] != same.user_nickname:
         try:
             nickname = User.objects.get(
@@ -78,8 +90,6 @@ def updateNickname(request):
 
 
 # 회원 페이지 이동 및 회원정보 가져오기
-
-
 def memberPage(request):
     # 로그인 했는지 확인
     if request.session.get('user_id'):
@@ -106,35 +116,43 @@ def memberPage(request):
 
         return render(request, 'member_app/memberPage.html', res_data)
     else:
-        return HttpResponseRedirect(reverse('signupPage'))
+        return HttpResponseRedirect(reverse('signinPage'))
 
 
 # 회원 탈퇴
 def deleteUser(request):
     if request.session.get('user_id'):
+        if request.method == 'POST':
+            checkpw = request.POST['checkpw']
+            print(checkpw)
+            user_session = request.session.get('user_id')
+            deluser = User.objects.get(u_id=user_session)
+            # 입력된 패스워드와 기존 유저의 패스워드가 같을 경우 삭제
+            if checkpw == deluser.user_pw:
+                meets = apps.get_model(
+                    app_label='noticeboard_app', model_name='meetings')
+                join = apps.get_model(
+                    app_label='joinus_app', model_name='joinus')
+                # 생성한 모임들 삭제
+                delmeet = meets.objects.filter(m_manager_id=user_session)
+                for m in delmeet:  # join에 등록된 모임도 삭제
+                    deljoin = join.objects.filter(m_id=m.m_id)
+                    deljoin.delete()
 
-        checkpw = request.POST['checkpw']
-        print(checkpw)
-        user_session = request.session.get('user_id')
-        deluser = User.objects.get(u_id=user_session)
-        if checkpw == deluser.user_pw:
-            meets = apps.get_model(
-                app_label='noticeboard_app', model_name='meetings')
-            join = apps.get_model(app_label='joinus_app', model_name='joinus')
-            # 생성한 모임 삭제
-            delmeet = meets.objects.filter(m_manager_id=user_session)
-            for m in delmeet:  # join에 등록된 모임도 삭제
-                deljoin = join.objects.filter(m_id=m.m_id)
+                delmeet.delete()
+                # 가입한 모임 삭제
+                deljoin = join.objects.filter(u_id=user_session)
                 deljoin.delete()
+                # 유저 정보 삭제
+                deluser.delete()
+                # 세션에 등록된 유저 정보도 삭제
+                request.session.clear()
+                auth.logout(request)
+                return HttpResponseRedirect(reverse('signinPage'))
+            else:
+                # 비번이 일치하지 않을 경우 넘기기
+                return HttpResponseRedirect(reverse('memberPage'))
 
-            delmeet.delete()
-            # 가입한 모임 삭제
-            deljoin = join.objects.filter(u_id=user_session)
-            deljoin.delete()
-            # 유저 정보 삭제
-            deluser.delete()
-            request.session.clear()
-        else:
-            # 비번이 일치하지 않을 경우 넘기기
+        elif request.method == 'GET':
             return HttpResponseRedirect(reverse('memberPage'))
     return HttpResponseRedirect(reverse('index'))
